// src/types/coreObjective.ts
export type CoreObjective = {
  generalCondition: string;
  consciousness: string;
  gcsEye: string;
  gcsVerbal: string;
  gcsMotor: string;
  systolic: string;
  diastolic: string;
  pulse: string;
  respiratoryRate: string;
  temperature: string;
  spo2: string;
  oxygenVia: string;
  weight: string;
  height: string;
  bmi: string;
  nutritionStatus: string;
  headNeck: string;
  thorax: string;
  abdomen: string;
  extremities: string;
  painNrs: string;
  otherFindings: string;
};

export const emptyCoreObjective: CoreObjective = {
  generalCondition: "",
  consciousness: "",
  gcsEye: "",
  gcsVerbal: "",
  gcsMotor: "",
  systolic: "",
  diastolic: "",
  pulse: "",
  respiratoryRate: "",
  temperature: "",
  spo2: "",
  oxygenVia: "",
  weight: "",
  height: "",
  bmi: "",
  nutritionStatus: "",
  headNeck: "",
  thorax: "",
  abdomen: "",
  extremities: "",
  painNrs: "",
  otherFindings: "",
};